% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{arrange_isobole}
\alias{arrange_isobole}
\title{Arrange the isobole more robustly}
\usage{
arrange_isobole(iso_unordered, yps = 1e-09)
}
\arguments{
\item{iso_unordered}{unordered data.table(xp,yp)}

\item{yps}{an offset for zero-values}
}
\value{
an ordered isobole
}
\description{
Call arrange_isobole_internal in multiple ways to be a bit more robust.
Use either different metric or a different scale.
For all scale trafos, reverse them at the end to get the original isobole back
}
\details{
1. normal from x = 0 to y = 0
2. log scale
3. order from y = 0 to x = 0
4. log scale from y = 0 to x = 0
5. normal, L1 metric
6. normal, canberra metrix (see \code{\link[stats]{dist}})

Then, take the path with minimal path length, see \code{\link{getPathlength}}
}
