% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{determineRecalcNeccessary}
\alias{determineRecalcNeccessary}
\title{Determine if calculated quantities need to be updated}
\usage{
determineRecalcNeccessary(
  .inputFolder,
  fileToRecalc = file.path(.inputFolder, "GLOBAL-101-percentages-grid_fine.rds")
)
}
\arguments{
\item{.inputFolder}{Path to "Simulations" with simulation results}

\item{fileToRecalc}{e.g. "GLOBAL-101-percentages-grid_fine.rds"}
}
\description{
Check if some simulations are more recent than the file to recalculate.
This is useful for calculations which take some time, e.g. the 2d-quantiles ("percentages")
}
\examples{
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
populationIsoboles:::determineRecalcNeccessary(.inputFolder)
}
