% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{.axlab}
\alias{.axlab}
\title{Make nice axis label}
\usage{
.axlab(Compound)
}
\arguments{
\item{Compound}{list with compound info.}
}
\value{
a nice label
}
\description{
Make nice axis label
}
\examples{
populationIsoboles:::.axlab(list(Name = "Drug 1"))
}
