% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{.plotDocumentation}
\alias{.plotDocumentation}
\title{Documentation for plotting functions}
\usage{
.plotDocumentation(
  finalPaths,
  grid,
  allPaths,
  itermax_dt,
  grid_fine,
  levels,
  Compound1 = getCompoundInfo("Compound1"),
  Compound2 = getCompoundInfo("Compound2"),
  filename = NULL,
  .inputFolder,
  FLAGoverwriteQuantiles = determineRecalcNeccessary(.inputFolder),
  ...
)
}
\arguments{
\item{finalPaths}{Result from \link{PI_getFinalPaths}}

\item{grid}{Result from \link{PI_readGridValues}}

\item{allPaths}{Result from \link{PI_readAllPaths}}

\item{itermax_dt}{Result from \link{PI_readItermax}}

\item{grid_fine}{Result from \link{calculateCLRS}}

\item{levels}{Vector of confidence levels, e.g. c(0.5,0.95). For plotRibbon, named vector with min and max, e.g. c(min = 0.025, max = 0.975)}

\item{Compound1, Compound2}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{filename}{Path to save plot}

\item{.inputFolder}{Folder where simulations are stored. \cr
Example: file.path(.outputFolder, "Simulations")}

\item{FLAGoverwriteQuantiles}{Redo the calculation of the quantiles?}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\description{
This function is only there to document parameters of plotting functions.
This is necessary because inheritParams can only inherit from a single function
and no plot uses all arguments of all plots
}
\examples{
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
plotSpaghetti(.inputFolder)
plotQuantiles(.inputFolder)
plotCLRS(.inputFolder)
}
