% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{ensure_cross_in_vec}
\alias{ensure_cross_in_vec}
\title{Ensure a number is in a vector}
\usage{
ensure_cross_in_vec(vec, cross)
}
\arguments{
\item{vec}{vector to test against}

\item{cross}{numbers which should be elements of vec}
}
\value{
numeric(length(cross))
}
\description{
If cross not present in vec, choose the next bigger number of vec
}
\examples{
populationIsoboles:::ensure_cross_in_vec(vec = 1:10, cross = c(1,5,6.5))
}
