% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{extract_popkFromFilename}
\alias{extract_popkFromFilename}
\title{Get the identifier of a population from the filenam}
\usage{
extract_popkFromFilename(files)
}
\arguments{
\item{files}{vector of filenames}
}
\value{
Numeric vector of k's
}
\description{
Get the identifier of a population from the filenam
}
\examples{
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
files <- list.files(.inputFolder)
files <- grep("^(\\\\d+)[._]", files, value = TRUE)
try(populationIsoboles:::extract_popkFromFilename(files))
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
