% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{get_area}
\alias{get_area}
\title{Get area enclosed by isobole and right part of the grid}
\usage{
get_area(path, i, gridmin, gridmax, gridlens, imax)
}
\arguments{
\item{path}{data.table(xp,yp), e.g. output from \code{\link[=get_isobole_path]{get_isobole_path()}}}

\item{i}{Current iteration to subset gridlens}

\item{gridmin, gridmax, gridlens, imax}{See \code{\link[=grid_refinement_iteration]{grid_refinement_iteration()}} or \code{\link[=gridInfo_default]{gridInfo_default()}}}
}
\value{
A number between 0 and 1
}
\description{
This function calculates the fraction of the grid which is on the right hand side of the isobole, i.e.
outside of the polygon enclosed by the isobole and the coordinate axes.
}
\examples{
# Only approximate, gets better with higher imax
path <- data.table(xp = c(0,0.5,0.5), yp = c(0.5,.5,0))
imax <- 10
gridInfo <- populationIsoboles:::gridInfo_default(1,1,imax = imax)
i <- 9
do.call(populationIsoboles:::get_area, c(list(path =path, imax = imax, i = i), gridInfo))
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
