% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{get_gridlens}
\alias{get_gridlens}
\title{Construct a list of grid lenghts}
\usage{
get_gridlens(gridmin, gridmax, imax = 10)
}
\arguments{
\item{gridmin, gridmax}{named vectors}

\item{imax}{Maximum number of iterations}
}
\value{
list of vectors with decreasing gridlengths in powers of two.
}
\description{
Construct a list of grid lenghts
}
\examples{
gridmin <- c(x = 0, y = 0) + 0.0001
gridmax <- c(x = 10*2^7, y = 0.1*2^7) + 0.0001
populationIsoboles:::get_gridlens(gridmin, gridmax)
}
\seealso{
Other Grid refinement functions for isoboles: 
\code{\link{arrange_isobole_internal}()},
\code{\link{fastIsoboles_iteration0}()},
\code{\link{fastIsoboles_iteration}()},
\code{\link{fastIsoboles}()},
\code{\link{get_isobole_path}()}
}
\author{
Daniel Lill, IntiQuan \email{daniel.lill@intiquan.com}
}
\concept{Grid refinement functions for isoboles}
