% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{get_isobole_path}
\alias{get_isobole_path}
\title{Interpolate grid values and run contourLines to obtain isobole path}
\usage{
get_isobole_path(grid, gridmin, gridmax, gridlen_i, objvalue = 0.95)
}
\arguments{
\item{grid}{data.table with colums (x,y,evaluatedCum, objvalueCum) and others}

\item{gridmin, gridmax}{General grid parameters: Vectors with names x and y c(x = 0, y = 0)}

\item{gridlen_i}{Grid lengths of iteration i: c(x = 40, y = 40)}

\item{objvalue}{Height of the countour line, for 95\% objvalue 95}
}
\value{
list with a data.table with columns (xp,yp), the nodes of the isobole path, and a logical variable outsideGrid to indicate if the isobole is out of the defined grid
}
\description{
Interpolate grid values and run contourLines to obtain isobole path
}
\seealso{
Other Grid refinement functions for isoboles: 
\code{\link{arrange_isobole_internal}()},
\code{\link{fastIsoboles_iteration0}()},
\code{\link{fastIsoboles_iteration}()},
\code{\link{fastIsoboles}()},
\code{\link{get_gridlens}()}
}
\author{
Daniel Lill, IntiQuan \email{daniel.lill@intiquan.com}
}
\concept{Grid refinement functions for isoboles}
