% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{getCompoundInfo}
\alias{getCompoundInfo}
\title{Try getting the default Compound info}
\usage{
getCompoundInfo(Compound = "Compound1")
}
\arguments{
\item{Compound}{Example: "Compund1"}
}
\value{
IQRmalarie Compound info
}
\description{
If not existing, write dummy Compound Info
}
