% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{getPathlength}
\alias{getPathlength}
\title{Calculate the isobole path length}
\usage{
getPathlength(iso, FLAGreturnDT = FALSE)
}
\arguments{
\item{iso}{data.table(xp,yp, ...)}

\item{FLAGreturnDT}{FALSE: return total path length (numeric)
TRUE: Return data.table(xp,yp, distToNext, cumDist, totDist, ...)}
}
\value{
Depends on FLAGreturnDT: TRUE: data.table, FALSE: numeric(1L)
}
\description{
Calculate the isobole path length
}
\examples{
iso_unordered <- data.frame(
xp = c(00, 10, 10, 20, 30, 40, 50, 50, 50, 50, 60, 60, 60, 50, 50, 50),
yp = c(80, 80, 70, 60, 60, 70, 70, 60, 50, 40, 30, 40, 20, 20, 10, 0)
)
set.seed(1)
iso_unordered <- iso_unordered[sample(1:nrow(iso_unordered), nrow(iso_unordered)),]
iso_unordered <- data.table(iso_unordered)
iso_ordered <- populationIsoboles:::arrange_isobole_internal(iso_unordered)
populationIsoboles:::getPathlength(iso_unordered)
populationIsoboles:::getPathlength(iso_ordered)
}
