% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{ggplot_PI}
\alias{ggplot_PI}
\title{ggplot_PI}
\usage{
ggplot_PI(data = NULL, mapping = aes())
}
\arguments{
\item{data, mapping}{see ?ggplot}
}
\value{
ggplot
}
\description{
ggplot_PI
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
