% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{grid_refinement_iteration}
\alias{grid_refinement_iteration}
\title{Do one iteration of the grid refinement algorithm}
\usage{
grid_refinement_iteration(
  gridmin,
  gridmax,
  gridlen_i,
  path_j,
  grid_j,
  i,
  objfun,
  objvalue = 0.95
)
}
\arguments{
\item{gridmin, gridmax}{General grid parameters: Vectors with names x and y c(x = 0, y = 0)}

\item{gridlen_i}{Grid lengths of iteration i: c(x = 40, y = 40)}

\item{path_j, grid_j}{Values from iteration i-1.
path_j: data.table with columns (xp,yp) with nodes of iteration j's approximation of the isobole path
grid_j: data.table with columns (x,y,evaluatedCum, objvalueCum, evaluated1, ..., evaluatedj, objvalue1,..., objvaluej)
The *Cum columns contain all cumulated grid point evaluations of all iterations.
evaluatedCum has a 1, if the objvalue has been calculated at that point, or a zero if this value has not been calculated
The *j columns contain the information about iteration j.
evaluated2 has a 1 for all grid points which were evaluated in iteration 2.}

\item{i}{Identifier for iteration i}

\item{objfun}{Function(x,y) returning a vector of the same length with the corresponding objvalue values.
Combined with objvalue, this could be called the objective function}

\item{objvalue}{The value for which the contour lines are calculated. E.g. 0.95}
}
\value{
list(grid, path), as specified in the arguments `grid_j` and `path_j`
}
\description{
Do one iteration of the grid refinement algorithm
}
\examples{
# see grid_refinement()
}
\seealso{
Other Grid refinement functions for isoboles: 
\code{\link{arrange_isobole_internal}()},
\code{\link{get_gridlens}()},
\code{\link{get_isobole_path}()},
\code{\link{grid_refinement_iteration0}()},
\code{\link{grid_refinement}()}
}
\author{
Daniel Lill, IntiQuan \email{daniel.lill@intiquan.com}
}
\concept{Grid refinement functions for isoboles}
