% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{gridInfo_default}
\alias{gridInfo_default}
\title{Generate the default characteristics of the simulation grid}
\usage{
gridInfo_default(AMTx1Max = 1000, AMTx2Max = 1000, imax = 7, offset = 1e-12)
}
\arguments{
\item{AMTx1Max, AMTx2Max}{maximum dose for drug 1 and 2}

\item{imax}{Maximum number of iterations}

\item{offset}{small delta to avoid doses with value 0, which can cause problems if DOSELEVEL is a covariate}
}
\value{
list(gridmin, gridmax, gridlens)
}
\description{
Generate the default characteristics of the simulation grid
}
\examples{
populationIsoboles:::gridInfo_default()
}
