% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{list.files.zipped}
\alias{list.files.zipped}
\title{This function is an extension base::list.files. In addition to searching
in .inputFolder, it will search in all xxxx.zip files found in this input
folder, where xxxx are four decimal digits.
Generate a character vector of filenames matching a pattern, where
these files may be stored in .inputFolder or inside xxxx.zip files
inside .inputFolder.}
\usage{
list.files.zipped(
  path,
  pattern = NULL,
  all.files = FALSE,
  full.names = FALSE,
  recursive = FALSE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE
)
}
\arguments{
\item{path, pattern, all.files, full.names, recursive, ignore.case, include.dirs, no..}{See [base::list.files()]}
}
\description{
Returns a character vector, like base::list.files. Matched files in
.zip files are denoted as
"path/to/file/xxxx.zip@#@matched.file".
}
\details{
The above format is understood by the function readRDS.zipped
}
\examples{
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
populationIsoboles:::list.files.zipped(.inputFolder, "", full.names = TRUE)
}
