% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{output_gridrefinement_results}
\alias{output_gridrefinement_results}
\title{Output the results of the grid-refinement algorithm in a standardized manner}
\usage{
output_gridrefinement_results(isobole, k, .outputFolder)
}
\arguments{
\item{isobole}{result of \code{\link[=grid_refinement]{grid_refinement()}}}

\item{k}{Identifier of the current population}

\item{.outputFolder}{output folder where the results are stored. Will save the results in a subdirectory called "Simulate}
}
\value{
Null. This function is called for its side-effect.
}
\description{
Output the results of the grid-refinement algorithm in a standardized manner
}
