% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{PI_files}
\alias{PI_files}
\title{Generate default output file paths}
\usage{
PI_files(.outputFolder, k)
}
\arguments{
\item{.outputFolder}{results will be saved in file.path(.outputFolder, "Simulations")}

\item{k}{Index of current population}
}
\value{
list of file paths
}
\description{
Generate default output file paths
}
\examples{
populationIsoboles:::PI_files("Output", 1)
}
