% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_plotInternalAlgoExample}
\alias{PI_plotInternalAlgoExample}
\title{Plot multiple plots to demonstrate the algorithm}
\usage{
PI_plotInternalAlgoExample(
  grid,
  allPaths,
  itermax_dt,
  Popk = 1,
  Compound1 = getCompoundInfo("Compound1"),
  Compound2 = getCompoundInfo("Compound2"),
  .outputFolder = NULL,
  ...
)
}
\arguments{
\item{grid}{Result from \link{PI_readGridValues}}

\item{allPaths}{Result from \link{PI_readAllPaths}}

\item{itermax_dt}{Result from \link{PI_readItermax}}

\item{Popk}{Integer denoting the population to plot}

\item{Compound1}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{Compound2}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{.outputFolder}{Folder to save the plots}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\value{
ggplot
}
\description{
Plot multiple plots to demonstrate the algorithm
}
\seealso{
Other Internal isobole plotting functions: 
\code{\link{PI_plotInternalMonoDistributions}()},
\code{\link{PI_plotInternalQuantiles}()},
\code{\link{PI_plotInternalSingleExample}()},
\code{\link{PI_plotInternalSpaghetti}()}
}
\concept{Internal isobole plotting functions}
