% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_plotInternalMonoDistributions}
\alias{PI_plotInternalMonoDistributions}
\title{Plot the distributions of the mono doses}
\usage{
PI_plotInternalMonoDistributions(grid_fine, filename = NULL, ...)
}
\arguments{
\item{grid_fine}{Result from \link{calculateCLRS}}

\item{filename}{Path to save plot}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\value{
ggplot
}
\description{
Plot the distributions of the mono doses
}
\seealso{
Other Internal isobole plotting functions: 
\code{\link{PI_plotInternalAlgoExample}()},
\code{\link{PI_plotInternalQuantiles}()},
\code{\link{PI_plotInternalSingleExample}()},
\code{\link{PI_plotInternalSpaghetti}()}
}
\concept{Internal isobole plotting functions}
