% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_plotInternalRibbon2}
\alias{PI_plotInternalRibbon2}
\title{Plot ribbon which is robust with respect to shape}
\usage{
PI_plotInternalRibbon2(
  grid_fine,
  levels = c(min = 0.025, max = 0.975),
  Compound1 = getCompoundInfo("Compound1"),
  Compound2 = getCompoundInfo("Compound2"),
  filename = NULL,
  ...
)
}
\arguments{
\item{grid_fine}{Result from \link{aggregateIsoboles}}

\item{levels}{Vector of confidence levels, e.g. c(0.5,0.95). For plotRibbon, named vector with min and max, e.g. c(min = 0.025, max = 0.975)}

\item{Compound1}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{Compound2}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{filename}{Path to save plot}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\value{
ggplot object
}
\description{
Deprecated: Plot ribbon of values which are compatible with 95% cure.
This is based on a confidence interval closed on both sides which is not
what is of interest...
}
