% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_plotInternalSingleExample}
\alias{PI_plotInternalSingleExample}
\title{Plot one population isobole only}
\usage{
PI_plotInternalSingleExample(
  grid,
  allPaths,
  Popk = 1,
  Compound1 = getCompoundInfo("Compound1"),
  Compound2 = getCompoundInfo("Compound2"),
  filename = NULL,
  ...
)
}
\arguments{
\item{grid}{Result from \link{PI_readGridValues}}

\item{allPaths}{Result from \link{PI_readAllPaths}}

\item{Popk}{Integer denoting the population to plot}

\item{Compound1}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{Compound2}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{filename}{Path to save plot}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\value{
ggplot
}
\description{
Plot one population isobole only
}
\seealso{
Other Internal isobole plotting functions: 
\code{\link{PI_plotInternalAlgoExample}()},
\code{\link{PI_plotInternalMonoDistributions}()},
\code{\link{PI_plotInternalQuantiles}()},
\code{\link{PI_plotInternalSpaghetti}()}
}
\concept{Internal isobole plotting functions}
