% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_plotInternalSpaghetti}
\alias{PI_plotInternalSpaghetti}
\title{Plot isobole spaghetti plot}
\usage{
PI_plotInternalSpaghetti(
  finalPaths,
  Compound1 = getCompoundInfo("Compound1"),
  Compound2 = getCompoundInfo("Compound2"),
  filename = NULL,
  ...
)
}
\arguments{
\item{finalPaths}{Result from \link{PI_getFinalPaths}}

\item{Compound1}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{Compound2}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{filename}{Path to save plot}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\value{
ggplot
}
\description{
Plot isobole spaghetti plot
}
\seealso{
Other Internal isobole plotting functions: 
\code{\link{PI_plotInternalAlgoExample}()},
\code{\link{PI_plotInternalMonoDistributions}()},
\code{\link{PI_plotInternalQuantiles}()},
\code{\link{PI_plotInternalSingleExample}()}
}
\concept{Internal isobole plotting functions}
