% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_quickTableItermax}
\alias{PI_quickTableItermax}
\title{Summarize the number of iterations needed across populations}
\usage{
PI_quickTableItermax(.inputFolder)
}
\arguments{
\item{.inputFolder}{path/to/Simulations}
}
\value{
data.table(Iterations = number of iterations, N = how many populations took that many iterations)
}
\description{
Summarize the number of iterations needed across populations
}
\examples{
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
populationIsoboles:::PI_quickTableItermax(.inputFolder)
}
