% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_readItermax}
\alias{PI_readItermax}
\alias{PI_readAllPaths}
\alias{PI_getFinalPaths}
\alias{PI_readGridLens}
\alias{PI_readGridInfo}
\alias{PI_readGridValues}
\alias{PI_readAreas}
\title{Functions to read simulation results}
\usage{
PI_readItermax(.inputFolder)

PI_readAllPaths(.inputFolder, itermax_dt = PI_readItermax(.inputFolder))

PI_getFinalPaths(allPaths)

PI_readGridLens(.inputFolder)

PI_readGridInfo(.inputFolder)

PI_readGridValues(
  .inputFolder,
  itermax_dt = PI_readItermax(.inputFolder),
  gridlens = PI_readGridLens(.inputFolder)
)

PI_readAreas(.inputFolder, itermax_dt = PI_readItermax(.inputFolder))
}
\arguments{
\item{.inputFolder}{Folder where the results are stored.
Is searched recursively}

\item{itermax_dt}{Output from \code{\link[=PI_readItermax]{PI_readItermax()}}}

\item{allPaths}{Output from \code{\link[=PI_readAllPaths]{PI_readAllPaths()}}}

\item{gridlens}{output of \link{PI_readGridLens}}
}
\value{
data.table(itermax, k) containing the maximum iteration for
each population
}
\description{
I moved the output of 001-gridInfo.rds from .outputFolder to .outputFolder/Simulations.
This function searches both folders to allow backwards compatibility
and that .inputFolder now means the "Simulations"-folder for all PI_read* functions
}
\examples{
library(populationIsoboles)
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
populationIsoboles:::PI_readItermax(.inputFolder)
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
populationIsoboles:::PI_readAllPaths(.inputFolder)
library(populationIsoboles)
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
populationIsoboles:::PI_readGridValues(.inputFolder)
}
\author{
Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com), Venelin Metiv (venelin.mitov@intiquan.com)
}
