% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{PI_readOutsideGrid}
\alias{PI_readOutsideGrid}
\title{Functions to read simulation results}
\usage{
PI_readOutsideGrid(.inputFolder)
}
\arguments{
\item{.inputFolder}{Folder where the results are stored. Is searched recursively}
}
\value{
data.table(outsideGrid) containing if the isobole is outside the defined grid for each population
}
\description{
Functions to read simulation results
}
\examples{
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
populationIsoboles:::PI_readOutsideGrid(.inputFolder)
}
