% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{PI_zip}
\alias{PI_zip}
\title{Zip files belonging to an isobole simulation}
\usage{
PI_zip(files)
}
\arguments{
\item{files}{List of filenames, result from \code{\link[=PI_files]{PI_files()}}}
}
\value{
NULL (called for side-effect)
}
\description{
Zip files belonging to an isobole simulation
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
