% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{plotCLRS}
\alias{plotCLRS}
\title{Plot confidence level response surface}
\usage{
plotCLRS(
  .inputFolder,
  filename = NULL,
  levels = c(0.5, 0.95),
  Compound1 = getCompoundInfo("Compound1"),
  Compound2 = getCompoundInfo("Compound2"),
  FLAGoverwriteQuantiles = determineRecalcNeccessary(.inputFolder = .inputFolder,
    fileToRecalc = "GLOBAL-101-percentages-grid_fine.rds"),
  ...
)
}
\arguments{
\item{.inputFolder}{Folder where simulations are stored. \cr
Example: file.path(.outputFolder, "Simulations")}

\item{filename}{Path to save plot}

\item{levels}{Vector of confidence levels, e.g. c(0.5,0.95). For plotRibbon, named vector with min and max, e.g. c(min = 0.025, max = 0.975)}

\item{Compound1}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{Compound2}{Character. Compound name, for cases when it's not available in the Malaria project}

\item{FLAGoverwriteQuantiles}{Redo the calculation of the quantiles?}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\description{
Plot confidence level response surface
}
\examples{
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
plotCLRS(.inputFolder)
}
\seealso{
Other Isobole plotting: 
\code{\link{plotRibbon}()}
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
\concept{Isobole plotting}
