% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{plotQuantiles}
\alias{plotQuantiles}
\title{Quick plotting of simulation results}
\usage{
plotQuantiles(
  .inputFolder,
  filename = NULL,
  FLAGoverwriteQuantiles = determineRecalcNeccessary(.inputFolder = .inputFolder,
    fileToRecalc = "GLOBAL-101-percentages-grid_fine.rds"),
  levels = c(0.025, 0.5, 0.975),
  ...
)
}
\arguments{
\item{.inputFolder}{Folder where simulations are stored. \cr
Example: file.path(.outputFolder, "Simulations")}

\item{filename}{Path to save plot}

\item{FLAGoverwriteQuantiles}{Redo the calculation of the quantiles?}

\item{levels}{Vector of confidence levels, e.g. c(0.5,0.95). For plotRibbon, named vector with min and max, e.g. c(min = 0.025, max = 0.975)}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\value{
ggplot
}
\description{
Quick plotting of simulation results
}
\seealso{
Other Isobole plotting
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
plotSpaghetti(.inputFolder): 
\code{\link{plotSpaghetti}()}
}
\concept{Isobole plotting
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
plotSpaghetti(.inputFolder)}
