% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{plotSpaghetti}
\alias{plotSpaghetti}
\title{Quick plotting of simulation results}
\usage{
plotSpaghetti(.inputFolder, filename = NULL, ...)
}
\arguments{
\item{.inputFolder}{Folder where simulations are stored. \cr
Example: file.path(.outputFolder, "Simulations")}

\item{filename}{Path to save plot}

\item{...}{Arguments goint to \code{\link[ggplot2]{ggsave}}}
}
\value{
A ggplot with the popIsoboles plotted as spaghetti
}
\description{
Quick plotting of simulation results
}
\seealso{
Other Isobole plotting
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
plotSpaghetti(.inputFolder): 
\code{\link{plotQuantiles}()}
}
\concept{Isobole plotting
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
plotSpaghetti(.inputFolder)}
