% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{readRDS.zipped}
\alias{readRDS.zipped}
\title{Reads an RDS file which can be stored in a xxxx.zip file
The filename argument specifies the path to the RDS file.
If the files is inside a zip file, then "@#@" is used as
a delimiter.}
\usage{
readRDS.zipped(filename)
}
\arguments{
\item{filename}{path to file}
}
\description{
Reads an RDS file which can be stored in a xxxx.zip file
The filename argument specifies the path to the RDS file.
If the files is inside a zip file, then "@#@" is used as
a delimiter.
}
\examples{
# Read parts from zipped files
.inputFolder <- system.file("examples/Example-1/Outputs/Simulations",
  package = "populationIsoboles")
files <- populationIsoboles:::list.files.zipped(.inputFolder, "", full.names = TRUE)
file <- grep("@#@", files, value =TRUE)[1]
print(file)
populationIsoboles:::readRDS.zipped(file)

# Read non-zipped files
tf <- tempfile(fileext = ".rds")
saveRDS(mtcars, tf)
populationIsoboles:::readRDS.zipped(tf)
}
