% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove points which are causing trouble}
\usage{
removeOutliers(iso)
}
\arguments{
\item{iso}{data.table(xp,yp) arranged with \link{arrange_isobole}}
}
\value{
data.table(xp,yp) with potential outliers removed
}
\description{
Sometimes arrange_isobole misses a point and then this point causes trouble
Identify them by finding points which have two large steps directly after each other
}
\examples{
# library(data.table)
# library(ggplot2)
#
# iso_outliers <- data.table(DATAPIOutlierExample1)
# iso_noOutliers <- removeOutliers(iso_outliers)
# PIggplot(mapping = aes(xp,yp)) +
# geom_path(data = iso_noOutliers, color = "black", linetype = 1) +
# geom_path(data = iso_outliers, color = "red", linetype = 2)
#
# iso_outliers <- data.table(DATAPIOutlierExample2)
# iso_noOutliers <- removeOutliers(iso_outliers)
# PIggplot(mapping = aes(xp,yp)) +
# geom_path(data = iso_noOutliers, color = "black", linetype = 1) +
# geom_path(data = iso_outliers, color = "red", linetype = 2)
}
