% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_populationIsoboles.R
\name{runPopulationIsoboles}
\alias{runPopulationIsoboles}
\title{Run population isoboles simulation}
\usage{
runPopulationIsoboles(
  Npop,
  Nsubj,
  objectiveValue,
  objectiveFunction,
  sampleIndividuals,
  argsSampleIndividuals,
  gridInfo = list(AMTx1Max = 1000, AMTx2Max = 1000),
  itermax = 5,
  FLAGverbose = FALSE,
  .outputFolder = "."
)
}
\arguments{
\item{Npop}{Number of populations}

\item{Nsubj}{Number of subjects per population}

\item{objectiveValue}{Target value}

\item{objectiveFunction}{Function to calculate target value. Needs to be function(AMTx1, AMTx2, parsIndiv),
where AMTx1, AMTx2 are numeric vectors and parsIndiv is the output of the supplied function \code{sampleIndividuals}}

\item{sampleIndividuals}{Function to sample individual parameters}

\item{argsSampleIndividuals}{List of arguments for \code{sampleIndividuals}}

\item{gridInfo}{A list(AMTx1Max, AMTx2Max), specifying the searched grid dimensions}

\item{itermax}{Number of iterations for fastIsoboles algorithm}

\item{FLAGverbose}{print messages to console?}

\item{.outputFolder}{Path to store results in}
}
\value{
Nothing, all results are written to disc
}
\description{
Run population isoboles simulation
}
\examples{
\dontrun{
# See examples in inst/examples/Example-01/01-RunExample.R
# See examples in inst/examples/Example-02/01-RunExample.R
}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
