% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{scale_fill_PI}
\alias{scale_fill_PI}
\title{scalefillPI
Copied from dMod}
\usage{
scale_fill_PI(...)
}
\arguments{
\item{...}{see scale_color_manual}
}
\value{
added to plots
}
\description{
scalefillPI
Copied from dMod
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
