% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_gridRefinementAlgorithm.R
\name{seqminmax}
\alias{seqminmax}
\title{seq including min and max}
\usage{
seqminmax(from, to, by)
}
\arguments{
\item{from, to, by}{as in \code{\link{seq}}}
}
\description{
seq including min and max
}
\examples{
seq(0,1,0.3)
populationIsoboles:::seqminmax(0,1,0.3)
}
\author{
Daniel Lill, IntiQuan \email{daniel.lill@intiquan.com}
}
