% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isobole_plotting.R
\name{theme_PI}
\alias{theme_PI}
\title{Nice ggplot theme}
\usage{
theme_PI(base_size = 12, base_family = "")
}
\arguments{
\item{base_size, base_family}{see ?theme_bw}
}
\description{
Taken from dMod
}
